+++ project/src/components/videoplayer/videoplayer.css
@@ -1,0 +1,5 @@
+.videoplayer {
+  width: 280px;
+  height: 175px;
+  object-fit: cover;
+}
